+++ iridauploader/core/upload_helpers.py
@@ -28,6 +28,11 @@ def _set_and_write_directory_status(directory_status, status, message=None):
     :param message: string
     :return:
     """
+    # If a run is being set to error, but it already has a sequencing id (started upload),
+    # it is a partial run that could be resumed.
+    if status == DirectoryStatus.ERROR and directory_status.run_id is not None:
+        status = DirectoryStatus.PARTIAL
+
     try:
         directory_status.status = status
         directory_status.message = message
@@ -76,8 +81,6 @@ def parse_and_validate(directory_status, parse_as_partial):
     :param parse_as_partial: sequencing_run will not include any samples that have already been uploaded
     :return:
     """
-    # Set directory status to partial before starting
-    _set_and_write_directory_status(directory_status, DirectoryStatus.PARTIAL)
 
     try:
         sequencing_run = parsing_handler.parse_and_validate(directory_status.directory)
@@ -225,10 +228,7 @@ def upload_sequencing_run(sequencing_run, directory_status, upload_mode, upload_
         logging.error("Lost connection to Irida")
         logging.error("Errors: " + pformat(e.args))
         full_error = "Lost connection to Irida. Errors: " + pformat(e.args)
-        if directory_status.run_id is not None:  # A run ID was generated, so the run is partial and can be continued
-            _set_and_write_directory_status(directory_status, DirectoryStatus.PARTIAL, full_error)
-        else:
-            _set_and_write_directory_status(directory_status, DirectoryStatus.ERROR, full_error)
+        _set_and_write_directory_status(directory_status, DirectoryStatus.ERROR, full_error)
         raise e
     except api.exceptions.IridaResourceError as e:
         logging.error("Could not access IRIDA resource")